bug fixed on plotting pages

+++ ECG/webEngine/ecg/ECG_reader.py
@@ -43,7 +43,7 @@ class ECG_reader():
         # original unit in dicom is milliVolt, converted unit in self.wavech is microVolt, assuming channel sensitivity's unit is mV
         self.wavech = [
                     [
-                        elem*self.sensitivity*1000 for index, elem in enumerate( wavedata ) if index % self.NumofChannels == channel_number
+                        int(elem*self.sensitivity*1000) for index, elem in enumerate( wavedata ) if index % self.NumofChannels == channel_number
                     ] for channel_number in range( self.NumofChannels )
                 ]    
         print 'there is ', len(self.wavech), ' channels in the test dicom'
